Improved JSON loading syntax
Minor improvement to file loading: see [PEP 343](https://peps.python.org/pep-0343/) - this syntax adds a built-in `try:`/`finally:` statement around the `json.load()` call, which ensures the file is closed. This is a more robust way to implement the file load, especially valuable for open source code which may be subject to more modifications in the future. Also see [this](https://thispointer.com/python-open-a-file-using-open-with-statement-benefits-explained-with-examples/) page for more practical examples motivating this syntax.